requirements.txt
pydantic>=2.0.0
openai>=1.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
pytest>=7.0.0
